conflicto solucionado y e implementacion agregada de mi ticket
