+++ src/main/java/com/alkemy/ong/services/UserService.java
@@ -6,6 +6,7 @@ import com.alkemy.ong.dto.UserDTORequest;
 import com.alkemy.ong.entities.User;
 import javassist.NotFoundException;
 import com.alkemy.ong.exception.AmazonS3Exception;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 
@@ -21,5 +22,6 @@ public interface UserService {
   String delete(String id) throws NotFoundException;
   UserDTO updateUser(MultipartFile file, UserDTORequest userDTOrequest) throws NotFoundException, IOException, AmazonS3Exception;
   List<UserDTO> getAll();
+  UserDTO getMe(String jwt) throws Exception;
 
 }
