+++ src/main/java/com/alkemy/ong/services/impl/UserServiceImpl.java
@@ -86,5 +86,15 @@ public class UserServiceImpl implements UserService {
         return usersDTO;
 
     }
+    @Autowired
+    JwtService jwtService;
+    public UserDTO getMe(String jwt) throws Exception{
+        String emailUser = jwtService.getUsername(jwt);
+        Boolean exitsUser = userRepo.existsByEmail(emailUser);
+        if (!exitsUser) throw new NotFoundException("A user with this token was not found");
+        Optional<User> user = userRepo.findByEmail(emailUser);
+        return mapper.userEntity2DTO(user.get());
+
+    }
 
 }
